# openapi.json
{
  "components": {
    "schemas": {
...
      "GeoStat": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "country_iso": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Iso"
          },
          "country_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Name"
          },
          "pct": {
            "title": "Pct",
            "type": "number"
          }
        },
        "required": [
          "country_iso",
          "country_name",
          "count",
          "pct"
        ],
        "title": "GeoStat",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginIn": {
        "properties": {
          "password": {
            "title": "Password",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginIn",
        "type": "object"
      },
      "OverviewStats": {
        "properties": {
          "avg_threat_score": {
            "title": "Avg Threat Score",
            "type": "number"
          },
          "period_days": {
            "title": "Period Days",
            "type": "integer"
          },
          "sessions_by_service": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Sessions By Service",
            "type": "object"
          },
          "severity_breakdown": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Severity Breakdown",
            "type": "object"
          },
          "top_countries": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Top Countries",
            "type": "array"
          },
          "top_ttps": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Top Ttps",
            "type": "array"
          },
          "total_sessions": {
            "title": "Total Sessions",
            "type": "integer"
          },
          "unique_ips": {
            "title": "Unique Ips",
            "type": "integer"
          }
        },
        "required": [
          "period_days",
          "total_sessions",
          "unique_ips",
          "sessions_by_service",
          "severity_breakdown",
          "top_countries",
          "top_ttps",
          "avg_threat_score"
        ],
        "title": "OverviewStats",
        "type": "object"
      },
...
      "TTPStat": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "pct": {
            "title": "Pct",
            "type": "number"
          },
          "tactic": {
            "title": "Tactic",
            "type": "string"
          },
          "technique_id": {
            "title": "Technique Id",
            "type": "string"
          }
        },
        "required": [
          "technique_id",
          "name",
          "tactic",
          "count",
          "pct"
        ],
        "title": "TTPStat",
        "type": "object"
      },
      "TimelineBucket": {
        "properties": {
          "avg_score": {
            "title": "Avg Score",
            "type": "number"
          },
          "bucket": {
            "format": "date-time",
            "title": "Bucket",
            "type": "string"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "bucket",
          "count",
          "avg_score"
        ],
        "title": "TimelineBucket",
        "type": "object"
      },
      "TokenOut": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in"
        ],
        "title": "TokenOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "HoneyStrike API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/login": {
      "post": {
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/refresh": {
      "post": {
        "operationId": "refresh_token_api_auth_refresh_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "hs_refresh",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hs Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Token",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/stats/geo": {
      "get": {
        "operationId": "geo_api_stats_geo_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 7,
              "maximum": 365,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GeoStat"
                  },
                  "title": "Response Geo Api Stats Geo Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Geo",
        "tags": [
          "stats"
        ]
      }
    },
    "/api/stats/overview": {
      "get": {
        "operationId": "overview_api_stats_overview_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 7,
              "maximum": 365,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewStats"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Overview",
        "tags": [
          "stats"
        ]
      }
    },
    "/api/stats/timeline": {
      "get": {
        "operationId": "timeline_api_stats_timeline_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 7,
              "maximum": 90,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "bucket",
            "required": false,
            "schema": {
              "default": "hour",
              "pattern": "^(hour|day)$",
              "title": "Bucket",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TimelineBucket"
                  },
                  "title": "Response Timeline Api Stats Timeline Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Timeline",
        "tags": [
          "stats"
        ]
      }
    },
    "/api/stats/ttps": {
      "get": {
        "operationId": "ttps_api_stats_ttps_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 365,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TTPStat"
                  },
                  "title": "Response Ttps Api Stats Ttps Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ttps",
        "tags": [
          "stats"
        ]
      }
    },
...
  }
}